+++ MediaToolkit src/MediaToolkit.Test/Features/IArgument.cs
@@ -1,0 +1,11 @@
+﻿using System.Text;
+
+namespace MediaToolkit.Test.Features
+{
+    public interface IArgument
+    {
+        StringBuilder Argument { get; }
+
+        void ComposeArgument();
+    }
+}
