+++ MediaToolkit src/MediaToolkit.Test/Features/IMediaProcessor.cs
@@ -1,0 +1,12 @@
+﻿namespace MediaToolkit.Test.Features
+{
+    /*
+    * This will eventually replace the "Engine" class
+    */
+
+    public interface IMediaProcessor<TArgument> where TArgument : IArgument
+    {
+
+
+    }
+}
