+++ MediaToolkit src/MediaToolkit/Features/IMetadata.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace MediaToolkit.Features
+{
+    public interface IMetadata
+    {
+        Dictionary<string, string> MetadataIndex { get; set; } 
+    }
+}
