+++ MediaToolkit src/MediaToolkit/Features/IMetadataProvider.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace MediaToolkit.Features
+{
+    public interface IMetadataProvider : IDisposable
+    {
+        IMetadata GetMetadata(string filename);
+    }
+}
