+++ MediaToolkit src/MediaToolkit/Features/Metadata.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace MediaToolkit.Features
+{
+    public class Metadata : IMetadata
+    {
+        public Dictionary<string, string> MetadataIndex { get; set; }
+    }
+}
